requirements.txt
Flask==0.10.1
Flask-SSLify==0.1.4
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
gunicorn==19.1.1
itsdangerous==0.24
pypng==0.0.17
virtualenv==1.11.6
